moved some stuff around, truncated the quote in teh drawer
