+++ app/lish/[did]/[publication]/[rkey]/PostHeader/CollapsedPostHeader.tsx
@@ -1,7 +1,10 @@
 "use client";
 
 import { Media } from "components/Media";
-import { Interactions, useInteractionState } from "../Interactions";
+import {
+  Interactions,
+  useInteractionState,
+} from "../Interactions/Interactions";
 import { useState, useEffect } from "react";
 
 export const CollapsedPostHeader = (props: { title: string }) => {
@@ -34,10 +37,11 @@ export const CollapsedPostHeader = (props: { title: string }) => {
       className="fixed top-0 left-0 right-0 w-full bg-[#FDFCFA] border-b border-border-light"
     >
       <div className="flex gap-2 items-center justify-between px-3 pt-2 pb-0.5 ">
-        <div className="text-tertiary font-bold text-sm truncate">
-          {props.title}{" "}
+        <div className="text-tertiary font-bold text-sm truncate pr-1">
+          {props.title}
         </div>
         <Interactions />
+        <div className="w-4 h-4 rounded-full bg-test shrink-0" />
       </div>
     </Media>
   );
